restructered project and changed DJ
